# apps/banking-api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Intergalactic Banking API
  version: 2.0.0
  description: A deterministic, run-scoped Banking API for comparing direct agent integrations with Postman
    Fabric Gateway. Monetary values are non-negative safe integers in minor units; all fictional currencies
    currently use exponent 0.
  contact:
    name: Intergalactic Banking API team
    url: https://www.postman.com/
  license:
    name: ISC
servers:
  - url: http://localhost:3000
    description: Local demo server
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/auth:
    get:
      tags:
        - Admin
      summary: Create a deprecated development credential
      description: Create a deprecated development credential in the selected run.
      operationId: generateLegacyApiKey
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
      responses:
        "200":
          description: Create a deprecated development credential succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/TooManyRequests"
      security: []
  /api/v1/admin/api-keys:
    post:
      tags:
        - Admin
      summary: Create an API credential
      description: Create an API credential in the selected run. Requires an ADMIN principal.
      operationId: createApiKey
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "201":
          description: Create an API credential succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
            Idempotency-Replayed:
              $ref: "#/components/headers/Idempotency-Replayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateCredentialRequest"
  /api/v1/accounts:
    get:
      tags:
        - Accounts
      summary: List owned accounts
      description: List owned accounts in the selected run.
      operationId: listAccounts
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: owner
          in: query
          required: false
          description: Optional owner filter.
          schema:
            type: string
        - name: createdAt
          in: query
          required: false
          description: Optional createdAt filter.
          schema:
            type: string
      responses:
        "200":
          description: List owned accounts succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    post:
      tags:
        - Accounts
      summary: Create an account
      description: Create an account in the selected run.
      operationId: createAccount
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "201":
          description: Create an account succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
            Idempotency-Replayed:
              $ref: "#/components/headers/Idempotency-Replayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
  /api/v1/accounts/{accountId}:
    get:
      tags:
        - Accounts
      summary: Get an owned account
      description: Get an owned account in the selected run.
      operationId: getAccount
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - name: accountId
          in: path
          required: true
          description: Run-scoped accountId value.
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Get an owned account succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    put:
      tags:
        - Accounts
      summary: Replace mutable account fields
      description: Replace mutable account fields in the selected run.
      operationId: updateAccount
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - name: accountId
          in: path
          required: true
          description: Run-scoped accountId value.
          schema:
            type: string
            minLength: 1
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Replace mutable account fields succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
            Idempotency-Replayed:
              $ref: "#/components/headers/Idempotency-Replayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountRequest"
    patch:
      tags:
        - Accounts
      summary: Update mutable account fields
      description: Update mutable account fields in the selected run.
      operationId: patchAccount
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - name: accountId
          in: path
          required: true
          description: Run-scoped accountId value.
          schema:
            type: string
            minLength: 1
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Update mutable account fields succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
            Idempotency-Replayed:
              $ref: "#/components/headers/Idempotency-Replayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountRequest"
    delete:
      tags:
        - Accounts
      summary: Deactivate an empty account
      description: Deactivate an empty account in the selected run.
      operationId: deleteAccount
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - name: accountId
          in: path
          required: true
          description: Run-scoped accountId value.
          schema:
            type: string
            minLength: 1
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Deactivate an empty account succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
            Idempotency-Replayed:
              $ref: "#/components/headers/Idempotency-Replayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /api/v1/demo/runs/{runId}/reset:
    post:
      tags:
        - Demo
      summary: Reset a canonical demo run
      description: Reset a canonical demo run in the selected run. Requires an ADMIN principal.
      operationId: resetDemoRun
      parameters:
        - $ref: "#/components/parameters/DemoRunId"
        - $ref: "#/components/parameters/RequestId"
        - name: runId
          in: path
          required: true
          description: Run-scoped runId value.
          schema:
            type: string
            minLength: 1
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Reset a canonical demo run succeeded.
          headers:
            X-Demo-Run-Id: *a1
            X-Request-Id: *a2
            X-RateLimit-Limit: *a3
            X-RateLimit-Remaining: *a4
            X-RateLimit-Reset: *a5
            Idempotency-Replayed:
              $ref: "#/components/headers/Idempotency-Replayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoRunResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: A credential resolving to a CUSTOMER or ADMIN principal.
  parameters:
    DemoRunId:
      name: X-Demo-Run-Id
      in: header
      required: false
      description: Isolates state for a comparison lane; defaults to default.
      schema:
        type: string
        pattern: ^[A-Za-z0-9_-]{1,64}$
        default: default
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      description: Caller-provided correlation identifier; generated deterministically when omitted.
      schema:
        type: string
        minLength: 1
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Required for mutations. Exact replays return the original sanitized response; conflicting or
        concurrent duplicates return 409.
      schema:
        type: string
        minLength: 1
        maxLength: 128
    Limit:
      name: limit
      in: query
      required: false
      description: Page size.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque collection-scoped continuation cursor.
      schema:
        type: string
  headers:
...
    Retry-After:
      description: Seconds until another request should be attempted.
      schema:
        type: integer
        minimum: 1
    Idempotency-Replayed:
      description: True when the response was replayed.
      schema:
        type: string
        const: "true"
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          required:
            - name
            - message
          properties:
            name:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
    Page:
      type: object
      additionalProperties: false
      required:
        - limit
        - nextCursor
        - hasMore
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 100
        nextCursor:
          type:
            - string
            - "null"
        hasMore:
          type: boolean
    Principal:
      type: object
      additionalProperties: false
      required:
        - principalId
        - role
      properties:
        principalId:
          type: string
          pattern: ^PRN-
        role:
          type: string
          enum:
            - CUSTOMER
            - ADMIN
        customerId:
          type: string
          pattern: ^CUS-
    CredentialCreated:
      type: object
      additionalProperties: false
      required:
        - apiKey
        - principal
      properties:
        apiKey:
          type: string
          writeOnly: true
        principal:
          $ref: "#/components/schemas/Principal"
    Account:
      type: object
      additionalProperties: false
      required:
        - accountId
        - owner
        - accountType
        - createdAt
        - balance
        - currency
        - status
      properties:
        accountId:
          type: string
        owner:
          type: string
        accountType:
          type: string
          enum:
            - STANDARD
            - PREMIUM
            - BUSINESS
        createdAt:
          type: string
          format: date
        balance: &a6
          type: integer
          minimum: 0
          maximum: 9007199254740991
        currency: &a7
          type: string
          enum:
            - COSMIC_COINS
            - GALAXY_GOLD
            - MOON_BUCKS
        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
...
    DemoRun:
      type: object
      additionalProperties: false
      required:
        - runId
        - seedVersion
        - accounts
        - transactions
        - disputes
      properties:
        runId:
          type: string
        seedVersion:
          const: fabric-banking-v2
        accounts:
          type: integer
        transactions:
          type: integer
        disputes:
          type: integer
    AccountResponse:
      type: object
      additionalProperties: false
      required:
        - account
      properties:
        account:
          $ref: "#/components/schemas/Account"
    AccountList:
      type: object
      additionalProperties: false
      required:
        - accounts
        - page
      properties:
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/Account"
        page:
          $ref: "#/components/schemas/Page"
...
    DemoRunResponse:
      type: object
      additionalProperties: false
      required:
        - run
      properties:
        run:
          $ref: "#/components/schemas/DemoRun"
    CreateAccountRequest:
      type: object
      additionalProperties: false
      required:
        - owner
        - currency
      properties:
        owner:
          type: string
          minLength: 1
        balance:
          type: integer
          minimum: 0
          maximum: 9007199254740991
        currency: *a8
        accountType:
          type: string
          enum:
            - STANDARD
            - PREMIUM
            - BUSINESS
    UpdateAccountRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        owner:
          type: string
          minLength: 1
        accountType:
          type: string
          enum:
            - STANDARD
            - PREMIUM
            - BUSINESS
...
    AdminCreateCredentialRequest:
      type: object
      additionalProperties: false
      required:
        - role
      if:
        properties:
          role:
            const: CUSTOMER
      then:
        required:
          - customerId
      properties:
        role:
          type: string
          enum:
            - CUSTOMER
            - ADMIN
        customerId:
          type: string
  responses:
    BadRequest:
      description: The request is malformed or violates a schema.
      headers: &a10
        X-Demo-Run-Id: *a1
        X-Request-Id: *a2
        X-RateLimit-Limit: *a3
        X-RateLimit-Remaining: *a4
        X-RateLimit-Reset: *a5
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: A valid API key is required.
      headers: *a10
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The principal lacks permission.
      headers: *a10
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The resource does not exist or is not visible.
      headers: *a10
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: A state transition, uniqueness rule, or idempotency reservation conflicted.
      headers: *a10
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: The run-and-credential rate limit was exceeded.
      headers:
        X-Demo-Run-Id: *a1
        X-Request-Id: *a2
        X-RateLimit-Limit: *a3
        X-RateLimit-Remaining: *a4
        X-RateLimit-Reset: *a5
        Retry-After:
          $ref: "#/components/headers/Retry-After"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServiceUnavailable:
      description: A bounded run or idempotency store reached capacity.
      headers: *a10
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"